optibot-assistant/requirements.txt
openai
python-dotenv
pyyaml
click
rich
packaging
